Scan plugins folder for submenus
