+++ lib/cli/commands/index.js
@@ -4,3 +4,4 @@ export * from './growify-command.js';
 export * from './historical-command.js';
 export * from './pipes-command.js';
 export * from './new-submenu-command.js';
+export * from './scan-for-menus-command.js';
