+++ lib/cli/setup-cli.js
@@ -95,6 +95,11 @@ export function factory(program) {
 		.option('-d, --directory [dir]', 'The directory where the output path will be relative to. Defaults to your plugins folder')
 		.action(commands.newSubmenu);
 
+	program
+		.command('scan-for-menus [folder]')
+		.description('Scans the given folder for any submenus and adds them to the config file. Uses your configured plugin folder by default')
+		.action(commands.scanForMenus);
+
 	// Command for generating the optimal pipe layout in a city.
 	program
 		.command('pipes <city>')
